// lib/schemas/startup.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * StartupSphere API
 * API documentation for StartupSphere, the 3D mapping platform for startup ecosystems.
 * OpenAPI spec version: 1.0
 */

export interface Startup {
  bookmarks: number;
  categories: string[];
  contactInfo: string;
  description: string;
  foundedDate: string;
  founderName: string;
  id: number;
  likes: number;
  locationLat: number;
  locationLng: number;
  locationName: string;
  logoUrl: string;
  name: string;
  views: number;
  websiteUrl: string;
}